Update for edimax dongle in a Pi3
Since I develop on my Stratux using a second dongle I found that if you plug in an additional wifi dongle into the RPi3 then that dongle will/might be seen as wlan0. If this dongle is an edimax wifi dongle and it is detected as wlan0 then the current file will not be able to start the hostapd service.

This script will now ask the system to give the MAC address of the device at wlan0 and if the system is a RPi2 or 3.

[http://macaddress.webwat.ch/vendor/Edimax_Technology_Co._Ltd.](http://macaddress.webwat.ch/vendor/Edimax_Technology_Co._Ltd.)

If the system is a RPi2/3 and there is an edimax at wlan0 we will now start the hostapd-edimax driver with the hostapd-edimax.conf


Note:

As of right now there are 3 MAC blocks used by Edimax. I have 3 Edimax Nano's and all of them use the "74:da:38" Mac Block so I am not checking for the other 2 MAC blocks at this time. This could be changed in the future. 

I am only guessing that all the Edimax Nano's are in the 74:da:38 range and the 2 others are other devices.