Check for and handle hostapd.user
This script now not only starts hostapd using the proper driver(standard/edimax). 

It now checks for the existance of the hostapd.user file and if found uses it. if not found creates it with the values from the existing hostapd.conf

I choose to only carry over the values from the hostaod.conf and not the edimax because they should be the same. 

If there is no user configurable values in hostapd.conf and no hostapd.user file then a default hostapd.user file is created.