Added traffic type prefix definitions to traffic-help.html
Added additional information to traffic-help.html to include the prefix
used when Show Traffic Source In Callsign is enabled within the
settings.

It was difficult to locate the definitions of each of the traffic type
prefixes, so I thought the help page would be a good location to define
these.

Additionally adjusted some spacing of the list elements to provide
easier readability on this help page.