HTML/JS Prep for WiFi Settings and StaticIP value fix
FIrst, I added the missing code to pull in the Static IP info the input field from the JSON string

Second and biggest:
This is the code to add the WiFi Settings to the Settings Page. In preparation for the GO code behind the scenes

Before it is live/working a few lines should be removed and the modals cleaned up. 

I had some issues getting the ng-bind working in the modalErrorWiFi modal. There is an update issue. Something about being outside of the digest.  I'm not sure.

The code sends a JSON string:  {"WiFiSSID":"stratux","WiFiSecurityEnabled":true,"WiFiPasscode":"h98tdjepi","WiFiChannel":"5"} to /setSettings

I also did some cleanup in the HTML.