Allow updates from /boot/StratuxUpdates
I have often flown with a backup SD card. This SD card is a "virgin" card that has never been in my stratux before. I just write the latest IMG. 

I have always wanted to be able to have stratux update to the latest updates.sh files because there are features in the updates that are proven, and sometimes needed.  This would not be possible in flight with no internet and too distracting to update from my tablet in flight. 

My Idea was to allow update files to be placed on the /boot partition since it is visible to Windows. 

If you place an update, like update-stratux-v1.3b3-1877e66056.sh from http://updates.stratux.me/builds/, in a folder in /boot called StratuxUpdates(/boot/StratuxUpdates) it will automatically update that SH