ARP Table Check For Slow Clients
Updated isSleeping to check the Pi's ARP table for a client's MAC address.  If a client's MAC address is present in the ARP table *and* the client has responded to an ICMP ping request in the past 30 seconds, the client is marked as not sleeping.

Additionally, added logging statements to record when a client transitions from awake to sleep.