ffMonitor Updates Client Status
Updated the ffMonitor function to look for "i-can-play-ffm-udp" and "i-cannot-play-ffm-udp" from a client.  If either of those broadcast packets are received, we mark the client as awake.

This will help with older iOS hardware not responding via ICMP when their CPU gets busy.  iOS appears to stop replying to ICMP with high CPU, however packet captures have shown that the ForeFlight heartbeat broadcast packets are still sent during the high CPU times.